+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ActionsScreen.kt
@@ -86,9 +86,7 @@ fun ActionsScreen(modifier: Modifier = Modifier, viewModel: ConfigActionsViewMod
         )
     }
 
-    EnableFlashlightActionBottomSheet(viewModel.createActionDelegate)
-    ChangeFlashlightStrengthActionBottomSheet(viewModel.createActionDelegate)
-    HttpRequestBottomSheet(viewModel.createActionDelegate)
+    HandleActionBottomSheets(viewModel.createActionDelegate)
 
     ActionsScreen(
         modifier = modifier,