+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ChooseActionScreen.kt
@@ -50,6 +50,14 @@ import io.github.sds100.keymapper.common.utils.State
 import kotlinx.coroutines.flow.update
 
 @Composable
+fun HandleActionBottomSheets(delegate: CreateActionDelegate) {
+    EnableFlashlightActionBottomSheet(delegate)
+    ChangeFlashlightStrengthActionBottomSheet(delegate)
+    HttpRequestBottomSheet(delegate)
+    SmsActionBottomSheet(delegate)
+}
+
+@Composable
 fun ChooseActionScreen(
     modifier: Modifier = Modifier,
     viewModel: ChooseActionViewModel,
@@ -57,9 +65,7 @@ fun ChooseActionScreen(
     val state by viewModel.groups.collectAsStateWithLifecycle()
     val query by viewModel.searchQuery.collectAsStateWithLifecycle()
 
-    EnableFlashlightActionBottomSheet(viewModel.createActionDelegate)
-    ChangeFlashlightStrengthActionBottomSheet(viewModel.createActionDelegate)
-    HttpRequestBottomSheet(viewModel.createActionDelegate)
+    HandleActionBottomSheets(viewModel.createActionDelegate)
 
     ChooseActionScreen(
         modifier = modifier,
@@ -119,7 +125,7 @@ private fun ChooseActionScreen(
                     end = endPadding,
                 ),
 
-        ) {
+            ) {
             when (state) {
                 State.Loading -> LoadingScreen(modifier = Modifier.fillMaxSize())
 
@@ -224,7 +230,7 @@ private fun PreviewList() {
                                 icon = ComposeIconInfo.Vector(Icons.Rounded.Android),
                             ),
 
-                        ),
+                            ),
                     ),
                     SimpleListItemGroup(
                         header = "Connectivity",
@@ -245,7 +251,7 @@ private fun PreviewList() {
                                 isEnabled = false,
                             ),
 
-                        ),
+                            ),
                     ),
                 ),
             ),
@@ -274,7 +280,7 @@ private fun PreviewGrid() {
                                 icon = ComposeIconInfo.Vector(Icons.Rounded.Android),
                             ),
 
-                        ),
+                            ),
                     ),
                     SimpleListItemGroup(
                         header = "Connectivity",
@@ -304,10 +310,10 @@ private fun PreviewGrid() {
                                 isEnabled = false,
                             ),
 
-                        ),
+                            ),
                     ),
 
-                ),
+                    ),
             ),
         )
     }