+++ common/src/main/java/io/github/sds100/keymapper/common/utils/KMResult.kt
@@ -61,6 +61,8 @@ abstract class KMError : KMResult<Nothing>() {
     data object EnableImeFailed : KMError()
     data object NoAppToOpenUrl : KMError()
     data object NoAppToPhoneCall : KMError()
+    data object NoAppToSendSms : KMError()
+    data class SendSmsError(val resultCode: Int) : KMError()
 
     data class NotAFile(val uri: String) : KMError()
     data class NotADirectory(val uri: String) : KMError()