+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - #1773 Option to show floating buttons on top of keyboard or notification panel.
 - #1335 Intent API to enable/disable/toggle a key map.
 - #114 action to force stop app, and an action to clear an app from recents
+- #727 Actions to send SMS messages: "Send SMS" and "Compose SMS"
 
 ## Removed